# Covenant/API/Models/HostedFile.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace EasyPeasy.API.Models
{
    using Newtonsoft.Json;
    using System.Linq;

    public partial class HostedFile
    {
        /// <summary>
        /// Initializes a new instance of the HostedFile class.
        /// </summary>
        public HostedFile()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the HostedFile class.
        /// </summary>
        public HostedFile(int? id = default(int?), int? listenerId = default(int?), string path = default(string), string content = default(string))
        {
            Id = id;
            ListenerId = listenerId;
            Path = path;
            Content = content;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public int? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "listenerId")]
        public int? ListenerId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "path")]
        public string Path { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "content")]
        public string Content { get; set; }

    }
}