# Covenant/API/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "EasyPeasy API",
    "version": "v0.1"
  },
  "paths": {
...
    "/api/listeners/types/{id}": {
      "get": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "GetListenerType",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ListenerType"
            }
          }
        }
      }
    },
    "/api/listeners": {
      "get": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "GetListeners",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Listener"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "EditListener",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Listener"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Listener"
            }
          }
        }
      }
    },
    "/api/listeners/{id}": {
      "get": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "GetListener",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Listener"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "DeleteListener",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/listeners/http/{id}": {
      "get": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "GetHttpListener",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpListener"
            }
          }
        }
      }
    },
    "/api/listeners/http": {
      "post": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "CreateHttpListener",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/HttpListener"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpListener"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "EditHttpListener",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/HttpListener"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpListener"
            }
          }
        }
      }
    },
    "/api/listeners/bridge/{id}": {
      "get": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "GetBridgeListener",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BridgeListener"
            }
          }
        }
      }
    },
    "/api/listeners/bridge": {
      "post": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "CreateBridgeListener",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BridgeListener"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BridgeListener"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "EditBridgeListener",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BridgeListener"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BridgeListener"
            }
          }
        }
      }
    },
    "/api/listeners/{id}/hostedfiles": {
      "get": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "GetHostedFiles",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/HostedFile"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "CreateHostedFile",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/HostedFile"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HostedFile"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "EditHostedFile",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/HostedFile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HostedFile"
            }
          }
        }
      }
    },
    "/api/listeners/{id}/hostedfiles/{hfid}": {
      "get": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "GetHostedFile",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "hfid",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HostedFile"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "DeleteHostedFile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "hfid",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ListenerType": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
...
    "ProfileType": {
      "enum": [
        "HTTP",
        "Bridge"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ProfileType",
        "modelAsString": false
      }
    },
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/ProfileType"
        },
        "messageTransform": {
          "type": "string"
        }
      }
    },
    "ListenerStatus": {
      "enum": [
        "Uninitialized",
        "Active",
        "Stopped"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ListenerStatus",
        "modelAsString": false
      }
    },
    "Listener": {
      "required": [
        "bindAddress",
        "bindPort",
        "connectAddresses",
        "connectPort",
        "description",
        "anotherid",
        "listenerTypeId",
        "name",
        "profileId",
        "status"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "anotherid": {
          "maxLength": 100,
          "minLength": 0,
          "pattern": "^[a-zA-Z0-9]*$",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "bindAddress": {
          "pattern": "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$",
          "type": "string"
        },
        "bindPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "connectAddresses": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "connectPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "profileId": {
          "format": "int32",
          "type": "integer"
        },
        "profile": {
          "$ref": "#/definitions/Profile"
        },
        "listenerTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "listenerType": {
          "$ref": "#/definitions/ListenerType"
        },
        "status": {
          "$ref": "#/definitions/ListenerStatus"
        },
        "covenantUrl": {
          "type": "string"
        },
        "covenantToken": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
    "HostedFile": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "listenerId": {
          "format": "int32",
          "type": "integer"
        },
        "path": {
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      }
    },
...
    "HttpListener": {
      "required": [
        "bindAddress",
        "bindPort",
        "connectAddresses",
        "connectPort",
        "description",
        "anotherid",
        "listenerTypeId",
        "name",
        "profileId",
        "status",
        "urls",
        "useSSL"
      ],
      "type": "object",
      "properties": {
        "useSSL": {
          "type": "boolean"
        },
        "sslCertificate": {
          "type": "string"
        },
        "sslCertificatePassword": {
          "type": "string"
        },
        "sslCertHash": {
          "type": "string"
        },
        "urls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "anotherid": {
          "maxLength": 100,
          "minLength": 0,
          "pattern": "^[a-zA-Z0-9]*$",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "bindAddress": {
          "pattern": "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$",
          "type": "string"
        },
        "bindPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "connectAddresses": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "connectPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "profileId": {
          "format": "int32",
          "type": "integer"
        },
        "profile": {
          "$ref": "#/definitions/Profile"
        },
        "listenerTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "listenerType": {
          "$ref": "#/definitions/ListenerType"
        },
        "status": {
          "$ref": "#/definitions/ListenerStatus"
        },
        "covenantUrl": {
          "type": "string"
        },
        "covenantToken": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "BridgeListener": {
      "required": [
        "bindAddress",
        "bindPort",
        "connectAddresses",
        "connectPort",
        "description",
        "anotherid",
        "listenerTypeId",
        "name",
        "profileId",
        "status"
      ],
      "type": "object",
      "properties": {
        "isBridgeConnected": {
          "type": "boolean"
        },
        "implantReadCode": {
          "type": "string"
        },
        "implantWriteCode": {
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "anotherid": {
          "maxLength": 100,
          "minLength": 0,
          "pattern": "^[a-zA-Z0-9]*$",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "bindAddress": {
          "pattern": "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$",
          "type": "string"
        },
        "bindPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "connectAddresses": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "connectPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "profileId": {
          "format": "int32",
          "type": "integer"
        },
        "profile": {
          "$ref": "#/definitions/Profile"
        },
        "listenerTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "listenerType": {
          "$ref": "#/definitions/ListenerType"
        },
        "status": {
          "$ref": "#/definitions/ListenerStatus"
        },
        "covenantUrl": {
          "type": "string"
        },
        "covenantToken": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}